avalon-py/pyproject.toml
[tool.poetry]
authors = ["TsumiNa <liu.chang.1865@gmail.com>"]
description = ""
license = "License :: OSI Approved :: Apache Software License"
name = "avalon"
version = "0.1.0"

[tool.poetry.dependencies]
grpcio = "^1.38.0"
python = "^3.6"
pyyaml = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
